fix 32 bit arch

+++ src/Niancat.jl
@@ -58,7 +58,7 @@ end
 
 # Catch all other events we don't care about.
 on_error(h::NiancatHandler, e::EventError) = println("on_error: $e")
-on_reply(h::NiancatHandler, id::Int64, event::Event) = nothing
+on_reply(h::NiancatHandler, id::Int, event::Event) = nothing
 on_event(h::NiancatHandler, event::Event) = nothing
 
 #
